refactor: add types dir
