+++ src/main/@types/cli.ts
@@ -1,0 +1,9 @@
+export enum CliPackageManager {
+  npm = 'npm',
+  yarn = 'yarn',
+}
+
+export type CliArgs = {
+  recipe: string;
+  packageManager: CliPackageManager;
+};
